+++ src/level.cpp
@@ -779,9 +779,8 @@ size_t lev_current_idx;
 size_t lev_current_start;
 
 map_format_e lev_format;
-bool lev_force_xnod;
 
-bool lev_overflows;
+bool lev_overflows = false;
 
 // objects of loaded level, and stuff we've built
 std::vector<vertex_t *> lev_vertices;
@@ -1368,63 +1367,6 @@ static void GetLinedefsHexen(void)
   }
 }
 
-static inline uint16_t VanillaSegDist(const seg_t *seg)
-{
-  double lx = seg->side ? seg->linedef->end->x : seg->linedef->start->x;
-  double ly = seg->side ? seg->linedef->end->y : seg->linedef->start->y;
-
-  // use the "true" starting coord (as stored in the wad)
-  double sx = round(seg->start->x);
-  double sy = round(seg->start->y);
-
-  return static_cast<uint16_t>(floor(hypot(sx - lx, sy - ly) + 0.5));
-}
-
-static inline short_angle_t VanillaSegAngle(const seg_t *seg)
-{
-  // compute the "true" delta
-  double dx = round(seg->end->x) - round(seg->start->x);
-  double dy = round(seg->end->y) - round(seg->start->y);
-
-  double angle = ComputeAngle(dx, dy);
-
-  if (angle < 0)
-  {
-    angle += 360.0;
-  }
-
-  short_angle_t result = short_angle_t(floor(angle * 65536.0 / 360.0 + 0.5));
-
-  if (lev_format != MapFormat_Doom)
-  {
-    return result;
-  }
-
-  // -Elf- ZokumBSP
-  // 1080 => Additive degrees stored in tag
-  // 1081 => Set to degrees stored in tag
-  // 1082 => Additive BAM stored in tag
-  // 1083 => Set to BAM stored in tag
-  if (seg->linedef->special == Special_RotateDegrees)
-  {
-    result += DegreesToShortBAM(static_cast<uint16_t>(seg->linedef->tag));
-  }
-  else if (seg->linedef->special == Special_RotateDegreesHard)
-  {
-    result = DegreesToShortBAM(static_cast<uint16_t>(seg->linedef->tag));
-  }
-  else if (seg->linedef->special == Special_RotateAngleT)
-  {
-    result += static_cast<short_angle_t>(seg->linedef->tag);
-  }
-  else if (seg->linedef->special == Special_RotateAngleTHard)
-  {
-    result = static_cast<short_angle_t>(seg->linedef->tag);
-  }
-
-  return result;
-}
-
 /* ----- UDMF reading routines ------------------------- */
 
 static constexpr uint32_t UDMF_THING = 1;
@@ -1771,611 +1713,49 @@ void ParseUDMF(void)
 
 /* ----- writing routines ------------------------------ */
 
-static void MarkOverflow(int flags)
+//
+// Check Limits
+//
+
+static void CheckBinaryFormatLimits(void)
 {
-  // flags are ignored
-  lev_overflows = true;
-}
-
-static void PutVertices(void)
-{
-  // this size is worst-case scenario
-  size_t size = lev_vertices.size() * sizeof(raw_vertex_t);
-  Lump_c *lump = CreateLevelLump("VERTEXES", size);
-
-  size_t count = 0;
-  for (size_t i = 0; i < lev_vertices.size(); i++)
-  {
-    raw_vertex_t raw;
-
-    const vertex_t *vert = lev_vertices[i];
-
-    if (vert->is_new)
-    {
-      continue;
-    }
-
-    raw.x = GetLittleEndian(static_cast<int16_t>(floor(vert->x)));
-    raw.y = GetLittleEndian(static_cast<int16_t>(floor(vert->y)));
-
-    lump->Write(&raw, sizeof(raw));
-
-    count++;
-  }
-
-  lump->Finish();
-
-  if (count != num_old_vert)
-  {
-    PrintLine(LOG_ERROR, "PutVertices miscounted (%zu != %zu)", count, num_old_vert);
-  }
-
-  if (count > 65534)
-  {
-    PrintLine(LOG_NORMAL, "FAILURE: Number of vertices has overflowed.");
-    MarkOverflow(LIMIT_VERTEXES);
-  }
-}
-
-static inline uint16_t VertexIndex16Bit(const vertex_t *v)
-{
-  if (v->is_new)
-  {
-    return static_cast<uint16_t>(v->index | 0x8000U);
-  }
-
-  return static_cast<uint16_t>(v->index);
-}
-
-static inline uint32_t VertexIndex_XNOD(const vertex_t *v)
-{
-  if (v->is_new)
-  {
-    return static_cast<uint32_t>(num_old_vert + v->index);
-  }
-
-  return static_cast<uint32_t>(v->index);
-}
-
-static void PutSegs(void)
-{
-  // this size is worst-case scenario
-  size_t size = lev_segs.size() * sizeof(raw_seg_t);
-
-  Lump_c *lump = CreateLevelLump("SEGS", size);
-
-  for (size_t i = 0; i < lev_segs.size(); i++)
-  {
-    raw_seg_t raw;
-
-    const seg_t *seg = lev_segs[i];
-
-    raw.start = GetLittleEndian(VertexIndex16Bit(seg->start));
-    raw.end = GetLittleEndian(VertexIndex16Bit(seg->end));
-    raw.angle = GetLittleEndian(VanillaSegAngle(seg));
-    raw.linedef = GetLittleEndian(static_cast<uint16_t>(seg->linedef->index));
-    raw.flip = GetLittleEndian(seg->side);
-    raw.dist = GetLittleEndian(VanillaSegDist(seg));
-
-    // -Elf- ZokumBSP
-    if ((seg->linedef->dont_render_back && seg->side) || (seg->linedef->dont_render_front && !seg->side))
-    {
-      raw = {};
-    }
-
-    lump->Write(&raw, sizeof(raw));
-
-    if (HAS_BIT(config.debug, DEBUG_BSP))
-    {
-      PrintLine(LOG_DEBUG, "[%s] %zu  Vert %04X->%04X  Line %04X %s  Angle %04X  (%1.1f,%1.1f) -> (%1.1f,%1.1f)", __func__,
-                seg->index, GetLittleEndian(raw.start), GetLittleEndian(raw.end), GetLittleEndian(raw.linedef),
-                seg->side ? "L" : "R", GetLittleEndian(raw.angle), seg->start->x, seg->start->y, seg->end->x, seg->end->y);
-    }
-  }
-
-  lump->Finish();
-
-  if (lev_segs.size() > 65534)
-  {
-    PrintLine(LOG_NORMAL, "FAILURE: Number of segs has overflowed.");
-    MarkOverflow(LIMIT_SEGS);
-  }
-}
-
-static void PutSubsecs(void)
-{
-  size_t size = lev_subsecs.size() * sizeof(raw_subsec_t);
-
-  Lump_c *lump = CreateLevelLump("SSECTORS", size);
-
-  for (size_t i = 0; i < lev_subsecs.size(); i++)
-  {
-    raw_subsec_t raw;
-
-    const subsec_t *sub = lev_subsecs[i];
-
-    raw.first = GetLittleEndian(static_cast<uint16_t>(sub->seg_list->index));
-    raw.num = GetLittleEndian(static_cast<uint16_t>(sub->seg_count));
-
-    lump->Write(&raw, sizeof(raw));
-
-    if (HAS_BIT(config.debug, DEBUG_BSP))
-    {
-      PrintLine(LOG_DEBUG, "[%s] %zu  First %04X  Num %04X", __func__, sub->index, GetLittleEndian(raw.first),
-                GetLittleEndian(raw.num));
-    }
-  }
-
-  if (lev_subsecs.size() > 32767)
-  {
-    PrintLine(LOG_NORMAL, "FAILURE: Number of subsectors has overflowed.");
-    MarkOverflow(LIMIT_SSECTORS);
-  }
-
-  lump->Finish();
-}
-
-static size_t node_cur_index;
-
-static void PutOneNode(node_t *node, Lump_c *lump)
-{
-  if (node->r.node)
-  {
-    PutOneNode(node->r.node, lump);
-  }
-
-  if (node->l.node)
-  {
-    PutOneNode(node->l.node, lump);
-  }
-
-  node->index = node_cur_index++;
-
-  raw_node_t raw;
-
-  // note that x/y/dx/dy are always integral in non-UDMF maps
-  raw.x = GetLittleEndian(static_cast<int16_t>(floor(node->x)));
-  raw.y = GetLittleEndian(static_cast<int16_t>(floor(node->y)));
-  raw.dx = GetLittleEndian(static_cast<int16_t>(floor(node->dx)));
-  raw.dy = GetLittleEndian(static_cast<int16_t>(floor(node->dy)));
-
-  raw.b1.minx = GetLittleEndian(static_cast<int16_t>(node->r.bounds.minx));
-  raw.b1.miny = GetLittleEndian(static_cast<int16_t>(node->r.bounds.miny));
-  raw.b1.maxx = GetLittleEndian(static_cast<int16_t>(node->r.bounds.maxx));
-  raw.b1.maxy = GetLittleEndian(static_cast<int16_t>(node->r.bounds.maxy));
-
-  raw.b2.minx = GetLittleEndian(static_cast<int16_t>(node->l.bounds.minx));
-  raw.b2.miny = GetLittleEndian(static_cast<int16_t>(node->l.bounds.miny));
-  raw.b2.maxx = GetLittleEndian(static_cast<int16_t>(node->l.bounds.maxx));
-  raw.b2.maxy = GetLittleEndian(static_cast<int16_t>(node->l.bounds.maxy));
-
-  if (node->r.node)
-  {
-    raw.right = GetLittleEndian(static_cast<uint16_t>(node->r.node->index));
-  }
-  else if (node->r.subsec)
-  {
-    raw.right = GetLittleEndian(static_cast<uint16_t>(node->r.subsec->index | 0x8000));
-  }
-  else
-  {
-    PrintLine(LOG_ERROR, "Bad right child in node %zu", node->index);
-  }
-
-  if (node->l.node)
-  {
-    raw.left = GetLittleEndian(static_cast<uint16_t>(node->l.node->index));
-  }
-  else if (node->l.subsec)
-  {
-    raw.left = GetLittleEndian(static_cast<uint16_t>(node->l.subsec->index | 0x8000));
-  }
-  else
-  {
-    PrintLine(LOG_ERROR, "Bad left child in node %zu", node->index);
-  }
-
-  lump->Write(&raw, sizeof(raw));
-
-  if (HAS_BIT(config.debug, DEBUG_BSP))
-  {
-    PrintLine(LOG_DEBUG, "[%s] %zu  Left %04X  Right %04X  (%1.1f,%1.1f) -> (%1.1f,%1.1f)", __func__, node->index,
-              GetLittleEndian(raw.left), GetLittleEndian(raw.right), node->x, node->y, node->x + node->dx, node->y + node->dy);
-  }
-}
-
-static void PutNodes(node_t *root)
-{
-  size_t struct_size = sizeof(raw_node_t);
-
-  // this can be bigger than the actual size, but never smaller
-  size_t max_size = (lev_nodes.size() + 1) * struct_size;
-
-  Lump_c *lump = CreateLevelLump("NODES", max_size);
-
-  node_cur_index = 0;
-
-  if (root != nullptr)
-  {
-    PutOneNode(root, lump);
-  }
-
-  lump->Finish();
-
-  if (node_cur_index != lev_nodes.size())
-  {
-    PrintLine(LOG_ERROR, "PutNodes miscounted (%zu != %zu)", node_cur_index, lev_nodes.size());
-  }
-
-  if (node_cur_index > 32767)
-  {
-    PrintLine(LOG_NORMAL, "FAILURE: Number of nodes has overflowed.");
-    MarkOverflow(LIMIT_NODES);
-  }
-}
-
-static void CheckLimits(void)
-{
-  // this could potentially be 65536, since there are no reserved values
-  // for sectors, but there may be source ports or tools treating 0xFFFF
-  // as a special value, so we are extra cautious here (and in some of
-  // the other checks below, like the vertex counts).
-  if (lev_sectors.size() > 65535)
+  if (lev_sectors.size() > LIMIT_SECTOR)
   {
     PrintLine(LOG_NORMAL, "FAILURE: Map has too many sectors.");
-    MarkOverflow(LIMIT_SECTORS);
+    lev_overflows = true;
   }
 
-  // the sidedef 0xFFFF is reserved to mean "no side" in DOOM map format
-  if (lev_sidedefs.size() > 65535)
+  if (lev_sidedefs.size() > LIMIT_SIDE)
   {
     PrintLine(LOG_NORMAL, "FAILURE: Map has too many sidedefs.");
-    MarkOverflow(LIMIT_SIDEDEFS);
+    lev_overflows = true;
   }
 
-  // the linedef 0xFFFF is reserved for minisegs in GL nodes
-  if (lev_linedefs.size() > 65535)
+  if (lev_linedefs.size() > LIMIT_LINE)
   {
     PrintLine(LOG_NORMAL, "FAILURE: Map has too many linedefs.");
-    MarkOverflow(LIMIT_LINEDEFS);
-  }
-
-  if (!(config.force_xnod || config.ssect_xgl3))
-  {
-    if (num_old_vert > 32767 || num_new_vert > 32767 || lev_segs.size() > 32767 || lev_nodes.size() > 32767)
-    {
-      PrintLine(LOG_NORMAL, "WARNING: Forcing XNOD format nodes due to overflows.");
-      config.total_warnings++;
-      lev_force_xnod = true;
-    }
+    lev_overflows = true;
   }
 }
 
-struct Compare_seg_pred
+bsp_type_t CheckFormatBSP(void)
 {
-  inline bool operator()(const seg_t *A, const seg_t *B) const
+  if (lev_vertices.size() > LIMIT_VERT)
   {
-    return A->index < B->index;
-  }
-};
-
-void SortSegs(void)
-{
-  // sort segs into ascending index
-  std::sort(lev_segs.begin(), lev_segs.end(), Compare_seg_pred());
-
-  // remove unwanted segs
-  while (lev_segs.size() > 0 && lev_segs.back()->index == SEG_IS_GARBAGE)
-  {
-    UtilFree(lev_segs.back());
-    lev_segs.pop_back();
-  }
-}
-
-/* ----- ZDoom format writing --------------------------- */
-
-static void PutXnodVertices(Lump_c *lump)
-{
-  size_t orgverts = GetLittleEndian(num_old_vert);
-  size_t newverts = GetLittleEndian(num_new_vert);
-
-  lump->Write(&orgverts, 4);
-  lump->Write(&newverts, 4);
-
-  size_t count = 0;
-  for (size_t i = 0; i < lev_vertices.size(); i++)
-  {
-    raw_xnod_vertex_t raw;
-
-    const vertex_t *vert = lev_vertices[i];
-
-    if (!vert->is_new)
-    {
-      continue;
-    }
-
-    raw.x = GetLittleEndian(static_cast<int32_t>(floor(vert->x * 65536.0)));
-    raw.y = GetLittleEndian(static_cast<int32_t>(floor(vert->y * 65536.0)));
-
-    lump->Write(&raw, sizeof(raw));
-
-    count++;
+    PrintLine(LOG_NORMAL, "WARNING: Vertex overflow. Forcing XNOD node format.");
+    config.total_warnings++;
+    return BSP_XNOD;
   }
 
-  if (count != num_new_vert)
+  if (lev_vertices.size() <= LIMIT_VERT
+      && (lev_segs.size() > LIMIT_SEG || lev_subsecs.size() > LIMIT_SUBSEC || lev_nodes.size() > LIMIT_NODE))
   {
-    PrintLine(LOG_ERROR, "PutZVertices miscounted (%zu != %zu)", count, num_new_vert);
-  }
-}
-
-static void PutXnodSubsecs(Lump_c *lump)
-{
-  size_t raw_num = GetLittleEndian(lev_subsecs.size());
-  lump->Write(&raw_num, 4);
-
-  size_t cur_seg_index = 0;
-  for (size_t i = 0; i < lev_subsecs.size(); i++)
-  {
-    const subsec_t *sub = lev_subsecs[i];
-
-    raw_num = GetLittleEndian(sub->seg_count);
-    lump->Write(&raw_num, 4);
-
-    // sanity check the seg index values
-    size_t count = 0;
-    for (const seg_t *seg = sub->seg_list; seg; seg = seg->next, cur_seg_index++)
-    {
-      if (cur_seg_index != seg->index)
-      {
-        PrintLine(LOG_ERROR, "PutZSubsecs: seg index mismatch in sub %zu (%zu != %zu)", i, cur_seg_index, seg->index);
-      }
-
-      count++;
-    }
-
-    if (count != sub->seg_count)
-    {
-      PrintLine(LOG_ERROR, "PutZSubsecs: miscounted segs in sub %zu (%zu != %zu)", i, count, sub->seg_count);
-    }
+    PrintLine(LOG_NORMAL, "WARNING: BSP overflow. Forcing DeepBSPV4 node format.");
+    config.total_warnings++;
+    return BSP_DEEPBSPV4;
   }
 
-  if (cur_seg_index != lev_segs.size())
-  {
-    PrintLine(LOG_ERROR, "PutZSubsecs miscounted segs (%zu != %zu)", cur_seg_index, lev_segs.size());
-  }
-}
-
-static void PutXnodSegs(Lump_c *lump)
-{
-  size_t raw_num = GetLittleEndian(lev_segs.size());
-  lump->Write(&raw_num, 4);
-
-  for (size_t i = 0; i < lev_segs.size(); i++)
-  {
-    const seg_t *seg = lev_segs[i];
-
-    if (seg->index != i)
-    {
-      PrintLine(LOG_ERROR, "PutZSegs: seg index mismatch (%zu != %zu)", seg->index, i);
-    }
-
-    raw_xnod_seg_t raw = {};
-
-    raw.start = GetLittleEndian(VertexIndex_XNOD(seg->start));
-    raw.end = GetLittleEndian(VertexIndex_XNOD(seg->end));
-    raw.linedef = GetLittleEndian(static_cast<uint16_t>(seg->linedef->index));
-    raw.side = seg->side;
-    lump->Write(&raw, sizeof(raw));
-  }
-}
-
-static void PutXgl2Segs(Lump_c *lump)
-{
-  size_t raw_num = GetLittleEndian(lev_segs.size());
-  lump->Write(&raw_num, 4);
-
-  for (size_t i = 0; i < lev_segs.size(); i++)
-  {
-    const seg_t *seg = lev_segs[i];
-
-    if (seg->index != i)
-    {
-      PrintLine(LOG_ERROR, "PutXGL3Segs: seg index mismatch (%zu != %zu)", seg->index, i);
-    }
-
-    raw_xgl2_seg_t raw = {};
-
-    raw.vertex = GetLittleEndian(VertexIndex_XNOD(seg->start));
-    raw.partner = GetLittleEndian(static_cast<uint16_t>(seg->partner ? seg->partner->index : NO_INDEX));
-    raw.linedef = GetLittleEndian(static_cast<uint16_t>(seg->linedef ? seg->linedef->index : NO_INDEX));
-    raw.side = seg->side;
-
-    lump->Write(&raw, sizeof(raw));
-
-    if (HAS_BIT(config.debug, DEBUG_BSP))
-    {
-      PrintLine(LOG_DEBUG, "[%s] SEG[%zu] v1=%d partner=%d line=%d side=%d", __func__, i, raw.vertex, raw.partner, raw.linedef,
-                raw.side);
-    }
-  }
-}
-
-static void PutOneXnodNode(Lump_c *lump, node_t *node, bool xgl3)
-{
-  raw_xgl3_node_t raw;
-
-  if (node->r.node)
-  {
-    PutOneXnodNode(lump, node->r.node, xgl3);
-  }
-
-  if (node->l.node)
-  {
-    PutOneXnodNode(lump, node->l.node, xgl3);
-  }
-
-  node->index = node_cur_index++;
-
-  if (xgl3)
-  {
-    int32_t x = GetLittleEndian(static_cast<int32_t>(floor(node->x * 65536.0)));
-    int32_t y = GetLittleEndian(static_cast<int32_t>(floor(node->y * 65536.0)));
-    int32_t dx = GetLittleEndian(static_cast<int32_t>(floor(node->dx * 65536.0)));
-    int32_t dy = GetLittleEndian(static_cast<int32_t>(floor(node->dy * 65536.0)));
-
-    lump->Write(&x, 4);
-    lump->Write(&y, 4);
-    lump->Write(&dx, 4);
-    lump->Write(&dy, 4);
-  }
-  else
-  {
-    raw.x = GetLittleEndian(static_cast<int16_t>(floor(node->x)));
-    raw.y = GetLittleEndian(static_cast<int16_t>(floor(node->y)));
-    raw.dx = GetLittleEndian(static_cast<int16_t>(floor(node->dx)));
-    raw.dy = GetLittleEndian(static_cast<int16_t>(floor(node->dy)));
-
-    lump->Write(&raw.x, 2);
-    lump->Write(&raw.y, 2);
-    lump->Write(&raw.dx, 2);
-    lump->Write(&raw.dy, 2);
-  }
-
-  raw.b1.minx = GetLittleEndian(static_cast<int16_t>(node->r.bounds.minx));
-  raw.b1.miny = GetLittleEndian(static_cast<int16_t>(node->r.bounds.miny));
-  raw.b1.maxx = GetLittleEndian(static_cast<int16_t>(node->r.bounds.maxx));
-  raw.b1.maxy = GetLittleEndian(static_cast<int16_t>(node->r.bounds.maxy));
-
-  raw.b2.minx = GetLittleEndian(static_cast<int16_t>(node->l.bounds.minx));
-  raw.b2.miny = GetLittleEndian(static_cast<int16_t>(node->l.bounds.miny));
-  raw.b2.maxx = GetLittleEndian(static_cast<int16_t>(node->l.bounds.maxx));
-  raw.b2.maxy = GetLittleEndian(static_cast<int16_t>(node->l.bounds.maxy));
-
-  lump->Write(&raw.b1, sizeof(raw.b1));
-  lump->Write(&raw.b2, sizeof(raw.b2));
-
-  if (node->r.node)
-  {
-    raw.right = GetLittleEndian(static_cast<uint32_t>(node->r.node->index));
-  }
-  else if (node->r.subsec)
-  {
-    raw.right = GetLittleEndian(static_cast<uint32_t>(node->r.subsec->index | 0x80000000U));
-  }
-  else
-  {
-    PrintLine(LOG_ERROR, "Bad right child in ZDoom node %zu", node->index);
-  }
-
-  if (node->l.node)
-  {
-    raw.left = GetLittleEndian(static_cast<uint32_t>(node->l.node->index));
-  }
-  else if (node->l.subsec)
-  {
-    raw.left = GetLittleEndian(static_cast<uint32_t>(node->l.subsec->index | 0x80000000U));
-  }
-  else
-  {
-    PrintLine(LOG_ERROR, "Bad left child in ZDoom node %zu", node->index);
-  }
-
-  lump->Write(&raw.right, 4);
-  lump->Write(&raw.left, 4);
-
-  if (HAS_BIT(config.debug, DEBUG_BSP))
-  {
-    PrintLine(LOG_DEBUG, "[%s] %zu  Left %08X  Right %08X  (%f,%f) -> (%f,%f)", __func__, node->index,
-              GetLittleEndian(raw.left), GetLittleEndian(raw.right), node->x, node->y, node->x + node->dx, node->y + node->dy);
-  }
-}
-
-static void PutXnodNodes(Lump_c *lump, node_t *root)
-{
-  size_t raw_num = GetLittleEndian(lev_nodes.size());
-  lump->Write(&raw_num, 4);
-
-  node_cur_index = 0;
-
-  if (root)
-  {
-    PutOneXnodNode(lump, root, false);
-  }
-
-  if (node_cur_index != lev_nodes.size())
-  {
-    PrintLine(LOG_ERROR, "PutZNodes miscounted (%zu != %zu)", node_cur_index, lev_nodes.size());
-  }
-}
-
-static void PutXgl3Nodes(Lump_c *lump, node_t *root)
-{
-  size_t raw_num = GetLittleEndian(lev_nodes.size());
-  lump->Write(&raw_num, 4);
-
-  node_cur_index = 0;
-
-  if (root)
-  {
-    PutOneXnodNode(lump, root, true);
-  }
-
-  if (node_cur_index != lev_nodes.size())
-  {
-    PrintLine(LOG_ERROR, "PutZNodes miscounted (%zu != %zu)", node_cur_index, lev_nodes.size());
-  }
-}
-
-static size_t CalcXnodNodesSize(void)
-{
-  // compute size of the ZDoom format nodes.
-  // it does not need to be exact, but it *does* need to be bigger
-  // (or equal) to the actual size of the lump.
-
-  size_t size = 32; // header + a bit extra
-
-  size += 8 + lev_vertices.size() * sizeof(raw_xnod_vertex_t);
-  size += 4 + lev_subsecs.size() * sizeof(raw_xnod_subsec_t);
-  size += 4 + lev_segs.size() * sizeof(raw_xnod_seg_t);
-  size += 4 + lev_nodes.size() * sizeof(raw_xnod_node_t);
-
-  return size;
-}
-
-void SaveXnodFormat(node_t *root_node)
-{
-  SortSegs();
-
-  size_t max_size = CalcXnodNodesSize();
-
-  Lump_c *lump = CreateLevelLump("NODES", max_size);
-
-  lump->Write(XNOD_MAGIC, 4);
-
-  PutXnodVertices(lump);
-  PutXnodSubsecs(lump);
-  PutXnodSegs(lump);
-  PutXnodNodes(lump, root_node);
-
-  lump->Finish();
-  lump = nullptr;
-}
-
-static void SaveXgl3Format(Lump_c *lump, node_t *root_node)
-{
-  SortSegs();
-
-  // WISH : compute a max_size
-
-  lump->Write(XGL3_MAGIC, 4);
-
-  PutXnodVertices(lump);
-  PutXnodSubsecs(lump);
-  PutXgl2Segs(lump);
-  PutXgl3Nodes(lump, root_node);
-
-  lump->Finish();
-  lump = nullptr;
+  return BSP_VANILLA;
 }
 
 /* ----- whole-level routines --------------------------- */
@@ -2479,7 +1859,7 @@ static void AddMissingLump(const char *name, const char *after)
   cur_wad->AddLump(name)->Finish();
 }
 
-build_result_e SaveLevel(node_t *root_node)
+build_result_e SaveBinaryFormatLevel(node_t *root_node)
 {
   // Note: root_node may be nullptr
 
@@ -2493,47 +1873,44 @@ build_result_e SaveLevel(node_t *root_node)
   AddMissingLump("REJECT", "SECTORS");
   AddMissingLump("BLOCKMAP", "REJECT");
 
-  // user preferences
-  lev_force_xnod |= config.force_xnod;
-
   // check for overflows...
-  // this sets the force_xxx vars if certain limits are breached
-  CheckLimits();
+  CheckBinaryFormatLimits();
 
-  /* --- Normal nodes --- */
-  if (config.ssect_xgl3 && num_real_lines > 0)
+  bsp_type_t level_type = CheckFormatBSP();
+  level_type = std::max(config.bsp_type, level_type);
+
+  switch (level_type)
   {
-    // leave SEGS empty
-    CreateLevelLump("SEGS")->Finish();
-    Lump_c *lump = CreateLevelLump("SSECTORS");
-    SaveXgl3Format(lump, root_node);
-    CreateLevelLump("NODES")->Finish();
-  }
-  else if (lev_force_xnod && num_real_lines > 0)
-  {
-    CreateLevelLump("SEGS")->Finish();
-    CreateLevelLump("SSECTORS")->Finish();
-    // remove all the mini-segs from subsectors
-    NormaliseBspTree();
-    SaveXnodFormat(root_node);
-  }
-  else
-  {
-    // remove all the mini-segs from subsectors
-    NormaliseBspTree();
-
-    // reduce vertex precision for classic DOOM nodes.
-    // some segs can become "degenerate" after this, and these
-    // are removed from subsectors.
-    RoundOffBspTree();
-
-    SortSegs();
-
-    PutVertices();
-
-    PutSegs();
-    PutSubsecs();
-    PutNodes(root_node);
+  case BSP_XGL3:
+    {
+      SaveFormat_Xgl3(root_node);
+      break;
+    }
+  case BSP_XGL2:
+    {
+      SaveFormat_Xgl2(root_node);
+      break;
+    }
+  case BSP_XGLN:
+    {
+      SaveFormat_Xgln(root_node);
+      break;
+    }
+  case BSP_XNOD:
+    {
+      SaveFormat_Xnod(root_node);
+      break;
+    }
+  case BSP_DEEPBSPV4:
+    {
+      SaveFormat_DeepBSPV4(root_node);
+      break;
+    }
+  case BSP_VANILLA:
+    {
+      SaveFormat_Vanilla(root_node);
+      break;
+    }
   }
 
   PutBlockmap();
@@ -2552,7 +1929,7 @@ build_result_e SaveLevel(node_t *root_node)
   return BUILD_OK;
 }
 
-build_result_e SaveUDMF(node_t *root_node)
+build_result_e SaveTextMapLevel(node_t *root_node)
 {
   cur_wad->BeginWrite();
 
@@ -2571,7 +1948,7 @@ build_result_e SaveUDMF(node_t *root_node)
   }
   else
   {
-    SaveXgl3Format(lump, root_node);
+    SaveFormat_Xgl3(root_node);
   }
 
   // -Elf-
@@ -2745,10 +2122,10 @@ build_result_e BuildLevel(size_t lev_idx, const char *filename)
     {
     case MapFormat_Doom:
     case MapFormat_Hexen:
-      ret = SaveLevel(root_node);
+      ret = SaveBinaryFormatLevel(root_node);
       break;
     case MapFormat_UDMF:
-      ret = SaveUDMF(root_node);
+      ret = SaveTextMapLevel(root_node);
       break;
     default:
       break;